refactor: 合併 AIScoreDisplay 組件為單一檔案
- 刪除 AIScoreDisplayIndex.tsx (wrapper)
- 刪除舊的 AIScoreDisplay.tsx
- 保留 AIScoreDisplayEnhanced.tsx 並重命名為 AIScoreDisplay.tsx
- 更新所有 import 路徑

現在只有一個智能組件，會自動判斷是否有詳細資料並顯示對應介面
